Added code which replicated July 1 failure
This code successfully produced the same symptoms as was experienced by the instrument on July 1 (launch day). It was replicated while the instrument was outdoors, and indicated on the GPS module that it had connected to a satellite. Note that no datalogging was used, so the SD card is not among the suspects anymore of the sudden stop of collecting data.